# UCS/ARCH: Add SVE memcpy