@shamisp 2020-11-25
Using `ucs_memcpy_relaxed` for SM makes sense. @yosefe is there any particular research why it is not used ? I would guess it will benefit x86 as well ?