@yosefe 2020-11-27
> Using `ucs_memcpy_relaxed` for SM makes sense. @yosefe is there any particular research why it is not used ? I would guess it will benefit x86 as well ?

AFAIR it showed inferior results on some architectures/sizes, @hoopoepg @dmitrygx  do you remember?