@dmitrygx 2020-11-27
> AFAIR it showed inferior results on some architectures/sizes, @hoopoepg @dmitrygx do you remember?

yes, there is a 2x degradation with using `rep movsb` on AMD for copy-in/copy-out to/from shared regions