@shamisp 2020-11-27
> we had an idea to have a separate UCS wrapper for `memcpy()` for shared regions to tune them on different platforms, e.g. disable for AMD for now

Makes sense to have stripped down version of memcpy for shared memory copy.